user error

+++ src/main/java/com/example/quizzproject/entity/BaseEntity.java
@@ -28,7 +28,7 @@ public class BaseEntity implements Serializable {
     @LastModifiedDate
     private Date updatedAt;
 
-    private Long createdBy;
+    private int createdBy;
 
     private Long updatedBy;
